requirements.txt
azure-ai-documentintelligence==1.0.0b4
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.35.0
azure-storage-blob==12.25.1
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
docopt==0.6.2
dotenv==0.9.9
idna==3.10
isodate==0.7.2
msrest==0.7.1
oauthlib==3.3.1
pipreqs==0.4.13
pycparser==2.22
python-dotenv==1.1.1
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
yarg==0.1.10
